make more modular; streamline update page
