+++ hawc/apps/riskofbias/templates/riskofbias/arob_edit.html
@@ -1,7 +1,18 @@
 {% extends 'assessment-rooted.html' %}
 
 {% block content %}
+    <div class="btn-group float-right">
+        <a class="btn btn-primary dropdown-toggle" data-toggle="dropdown">Actions</a>
+        <div class="dropdown-menu dropdown-menu-right">
+            <span class="dropdown-header">{{assessment.get_rob_name_display}} changes</span>
+            <a class="dropdown-item" href="{% url 'riskofbias:arob_text_update' assessment.pk %}">Update instructions</a>
+            <div class="dropdown-divider"></div>
+            <span class="dropdown-header">Domain changes</span>
+            <a class="dropdown-item" href="{% url 'riskofbias:robd_create' assessment.pk %}">New domain</a>
+        </div>
+    </div>
     <h1>Study evaluation requirements</h1>
+    {{assessment.rob_settings.help_text|safe}}
     <div id='robTable'></div>
 {% endblock %}
